pyproject.toml
[tool.poetry]
name = "graphql-sync-dataloaders"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "graphql_sync_dataloaders"}]

[tool.poetry.dependencies]
python = "^3.8"
graphql-core = "~3.2.0"

[tool.poetry.group.dev.dependencies]
Django = "^4.1.1"
strawberry-graphql = "^0.131.5"
graphene-django = "3.0.0b7"
pytest = "^7.1.3"
black = "^22.8.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s"
DJANGO_SETTINGS_MODULE = "tests.django.django_settings"
testpaths = ["tests/"]
markers = ["django"]
